==> scripts/1. Dohko.ipynb <==
import os

import numpy as np
import pandas as pd
from scipy.io import loadmat

diags = [
    { 
        'desc': 'Atrial Fibrilation', 
        'diag': 'AF', 
        'code': '164889003' 
    },
    {
        'desc': 'Sinus Rhythm',
        'diag': 'SR',
        'code': '426783006'
    }
]
diags_code = np.array(list(map(lambda x: x['code'], diags)))
path_folder = '../databases/'
available_databases = os.listdir(path_folder)
available_databases.remove('.gitkeep')
available_databases

recordings = []

for path in available_databases:
    path = f'../databases/{path}/'
    headers = list(filter(lambda x: os.path.splitext(x)[1] == '.hea', 
                          sorted(os.listdir(path))))  
    hea_file = open(path + headers[0]).read()
    freq = int(hea_file.split('\n')[0].split(' ')[2])
           
    for filename in headers:
        filename, extension = os.path.splitext(filename)
        record_path = path + filename + '.mat'
        header_path = path + filename + '.hea'
        
        if (not os.path.isfile(record_path) or 
            not os.path.isfile(header_path)):
            continue
            
        header_file = open(header_path, 'r').read()
        
        rec_diags = []
        for row in header_file.split('\n'):
            if not row.startswith('#Dx'):
                continue
            entries = row.split(': ')[1].split(',')
            for entry in entries:
                rec_diags.append(entry.strip())

        rec_diags = np.array(rec_diags)
        diags_code = np.array(diags_code)
        intersec = np.intersect1d(rec_diags, diags_code)
        
        # if not intersec.size == 1 and rec_diags == 1:
        if not intersec.size == 1:
            '''
             0 - not found diag
             1 - found only one diag in diags
             2 - found two in diags? af and sr
            '''
            continue
            
        leads, sampling = loadmat(record_path)['val'].shape
        
        if int(leads) != 12:
            continue
        
        database = record_path.split('/')[-2]
        filename = record_path.split('/')[-1]
        diagnose = list(filter(lambda x: x['code'] == intersec[0], diags))[0]['diag']
        total_diags = int(rec_diags.size)
        time = float(sampling/freq)
        
        recordings.append([database, 
                           filename, 
                           diagnose, 
                           total_diags, 
                           freq,
                           time])
    print(len(recordings), path)

df = pd.DataFrame(recordings, columns=['db', 
                                       'filename', 
                                       'diag', 
                                       'total_diags', 
                                       'db_freq',
                                       'time'])
df.head()
print(df.shape)

df = df.sample(frac=1, ignore_index=True, random_state=32)
df.head()

print(f"AF: {str(df[df.diag == 'AF'].shape[0])} rows")
print(f"SR: {str(df[df.diag == 'SR'].shape[0])} rows")
df.groupby(['diag', 'db']).size()

df.to_csv('../workdata/dohko/dohko.csv', index=False)
