==> scripts/2. Miro.ipynb <==
import os
import math

import numpy as np
import pandas as pd
from scipy.io import loadmat
from scipy.signal import cheby2, sosfilt, resample
from tqdm.notebook import tqdm


dohko = pd.read_csv('../workdata/dohko/dohko.csv')
print(dohko.shape)
dohko.head()

n_rows = dohko.shape[1]
dohko['data'] = [[] for _ in range(n_rows)]

new_freq = 100

for index, row in tqdm(dohko.iterrows()):
    db = row['db']
    filename = row['filename']
    db_freq = row['db_freq']
    path = f'../databases/{db}/{filename}'
    record_file = loadmat(path)['val']
    recording = np.array(record_file, dtype=np.float64)
    n_leads, rec_samples = recording.shape 
    for lead in range(0, n_leads):
        sos = cheby2(12, 20, [0.35, 70], 'bandpass', fs=db_freq, output='sos')
        recording[lead, :] = sosfilt(sos, recording[lead, :])
    rec_time_sec = rec_samples/db_freq
    new_rec_sampling = int(rec_time_sec*new_freq)
    recording = resample(recording, new_rec_sampling, axis=1)
    dohko.at[index, 'data'] = recording

for index, row in enumerate(np.array_split(dohko, 10)):
    row.to_pickle(f'../workdata/miro/part_{index}.pkl')

### Flush RAM

os._exit(0)
