==> scripts/3. Aioros.ipynb <==
import os

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import neurokit2 as nk
import seaborn as sns

### Aioros

df = pd.read_pickle(f'../workdata/miro/part_0.pkl')
df = df.groupby(['diag', 'db']).head(5)
df.groupby(['diag', 'db']).size()

### Aioros Large

df = pd.read_pickle(f'../workdata/miro/part_0.pkl')
df = df.append(pd.read_pickle(f'../workdata/miro/part_1.pkl'))
df = df.append(pd.read_pickle(f'../workdata/miro/part_2.pkl'))

df.groupby(['diag', 'db']).size()
print(df.shape)
df.head()

### Aioros CPSC2018

parts = os.listdir(f'../workdata/miro/')

df = pd.DataFrame()

for p in parts:
    df_p = pd.read_pickle(f'../workdata/miro/{p}')
    df_p = df_p[df_p['db'] == 'WFDB_CPSC2018']
    df = df.append(df_p)

df.head()

leads = ('I', 'II', 'III', 'aVR', 'aVL', 'aVF', 'V1', 'V2', 'V3', 'V4', 'V5', 'V6')
sns.set(rc={'figure.figsize':(11.7,8.27)})
df2 = pd.DataFrame(columns=['diag', 'data'])

PRESENTATION_MODE = False

if PRESENTATION_MODE:
    if not os.path.exists(f'../workdata/aioros/SR/'):
        os.makedirs(f'../workdata/aioros/SR/')
    if not os.path.exists(f'../workdata/aioros/AF/'):
        os.makedirs(f'../workdata/aioros/AF/')
        
rows = df.shape[0]

for i, row in df.iterrows():
    print(f'{i}/{rows}')
    try:
        ecg_signal = row['data'][-2] # V5
        _, rpeaks = nk.ecg_peaks(ecg_signal, sampling_rate=100)
        ecg_r_peaks = rpeaks['ECG_R_Peaks']
        ecg_r_peaks = list(map(lambda x: x - 20, ecg_r_peaks))
        ecg_size = len(ecg_signal)
        ecg_r_peaks.insert(0, 0)
        ecg_r_peaks.insert(ecg_size, ecg_size)

        n_peaks = len(ecg_r_peaks)
        max_interval_RR = 0
        i_RR = 0

        for i_peak, peak in enumerate(ecg_r_peaks):
            if i_peak == n_peaks - 1:
                continue           
            interval_size = ecg_r_peaks[i_peak+1] - ecg_r_peaks[i_peak]
            if interval_size > max_interval_RR:
                max_interval_RR = interval_size
                i_RR = i_peak

        boundies = ecg_r_peaks[i_RR:i_RR+2]
        diag = row["diag"]
        new_row = {
            'db': row['db'],
            'filename': row['filename'],
            'diag': diag,
            'data': row['data'][:, boundies[0]:boundies[1]]
        }
        
        if PRESENTATION_MODE:
            if not os.path.exists(f'../workdata/aioros/{diag}/sample_{i}/'):
                os.makedirs(f'../workdata/aioros/{diag}/sample_{i}/')
            plot = sns.lineplot(
            x=np.arange(0, len(ecg_signal), 1, dtype=int),
            y=ecg_signal)
            plot.axvline(boundies[0], color='red')
            plot.axvline(boundies[1], color='red')
            fig = plot.get_figure()
            fig.set_size_inches(25,15)
            fig.savefig(f"../workdata/aioros/{diag}/sample_{i}/sample_v5.png", facecolor='white', transparent=False, dpi=300) 
            fig.clf()
            for j, lead in enumerate(leads):
                plot = sns.lineplot(x=np.arange(0, boundies[1]-boundies[0], 1, dtype=int), 
                                    y=new_row['data'][j])
                fig = plot.get_figure()
                fig.savefig(f"../workdata/aioros/{diag}/sample_{i}/lead_{lead}.png", facecolor='white', transparent=False, dpi=300) 
                fig.clf()
        df2 = df2.append(new_row, ignore_index=True)
    except Exception as e:
        print(e)
    
df2.to_pickle(f'../workdata/aioros/aioros_cpsc2018.pkl')
