==> scripts/6. Aphrodite.ipynb <==
import os

os.environ["LD_LIBRARY_PATH"] = "/home/renan/MATLAB/R2021b/runtime/glnxa64/";

import matlab
import pandas as pd
import numpy as np
import seaborn as sbn
from tensorflow.keras.utils import to_categorical
from tensorflow.keras.callbacks import ModelCheckpoint, EarlyStopping
from tensorflow.keras.models import load_model, Sequential
from tensorflow.keras.layers import Dense, Dropout, Conv2D,Conv3D, MaxPooling3D, MaxPooling2D, Flatten, BatchNormalization, Conv1D, LeakyReLU, MaxPool1D
from tensorflow.keras.regularizers import l2, l1_l2
from tensorflow.keras.optimizers import RMSprop
from sklearn.preprocessing import StandardScaler, LabelEncoder
from sklearn.model_selection import train_test_split
from scipy.signal import resample

from myPackages import myDehankelization, myHankelization

sc = StandardScaler()
le = LabelEncoder()

### Caracteristicas Camus

base_df = pd.read_csv('../workdata/shura/experiment2/results.csv')
base_df = base_df[base_df['rank'] == 5]
X = base_df[['iter', 'fval', 'relfval', 'relstep', 'delta', 'rho', 'relerr']].to_numpy()
X = sc.fit_transform(X)

Y = base_df[['diagnostic']]
Y = Y['diagnostic'].to_numpy()
le.fit(Y)
Y = le.transform(Y)

### Caracteristicas Afrodite

files = os.listdir('../workdata/shura/experiment2/rank_5/')
n_files = len(files)

df = pd.DataFrame(columns=['id', 'data', 'label'])

for i, filename in enumerate(files):
    print(f"{round((i*100)/n_files, 2)}%")
    with open(f'../workdata/shura/experiment2/rank_5/{filename}', 'rb') as f:
        output = np.load(f, allow_pickle=True)
    output = np.array(output)
    
    I = np.array(output[0][0]).shape[0]
    J = np.array(output[0][1]).shape[0]
    N = I*2 - 1 if I == J else I * 2

    sources = np.empty((0, N), float)
    contrib_sources = np.empty((0, 12), float)
    my_dehankelization = myDehankelization.initialize()
    for i in range(0, len(output)):
        a = np.array(output[i][0])
        b = np.array(output[i][1]).T
        c = np.array(output[i][2]).T
        contrib_sources = np.append(contrib_sources, c, axis=0)
        mixing_matrix = np.dot(a, b)
        m_hankel = matlab.double(mixing_matrix.tolist(), size=mixing_matrix.shape)
        source = my_dehankelization.dehankelization(m_hankel)
        sources = np.append(sources, np.array(source).T, axis=0)
    my_dehankelization.terminate()
    label = base_df.loc[int(filename.split('.')[0])].diagnostic
    result = {'id': int(filename.split('.')[0]),
              'data': sources,
              'label': label}
    df = df.append(result, ignore_index=True)
df.to_pickle('../workdata/aphrodite/sources_china.pkl')

df = pd.read_pickle('../workdata/aphrodite/sources_china.pkl')

# Sources extracted
S = 5

# Get the min size source between the samples
N = int(np.mean(list(map(lambda x: x.shape[1], df.data)))) # 96

for i, row in df.iterrows():
    sources = np.empty((S, N), float)
    for j, source in enumerate(row.data):
        source = source.reshape(1, source.shape[0])
        sources[j] = resample(source, N, axis=1)
    df.at[i, 'data'] = sources
  
Y = df[['label']]
Y = Y['label'].to_numpy()
le.fit(Y)
Y = le.transform(Y)

n_rows, _ = df.shape

## 1 dim

X = np.empty((n_rows, S*N))
for i in range(0, n_rows):
    data = df.iloc[i]['data']
    data = data.reshape(S*N)
    X[i, :] = data
    
X_train, X_test, y_train, y_test = train_test_split(X, Y, test_size=0.33, random_state=42)

X_train = X_train.reshape([*X_train.shape[0:3], 1])
X_test = X_test.reshape([*X_test.shape[0:3], 1])

y_train = np.asarray(y_train).astype('float32').reshape((-1,1))
y_test = np.asarray(y_test).astype('float32').reshape((-1,1))

cnn_model = Sequential()

cnn_model.add(Conv1D(filters=32, kernel_size=(6,), padding='same', activation=LeakyReLU(alpha=0.001), input_shape = (X_train.shape[1],1)))

model.add(BatchNormalization())

cnn_model.add(Conv1D(filters=64, kernel_size=(6,), padding='same', activation=LeakyReLU(alpha=0.001), input_shape = (X_train.shape[1],1)))

model.add(BatchNormalization())

cnn_model.add(Conv1D(filters=128, kernel_size=(6,), padding='same', activation=LeakyReLU(alpha=0.001), input_shape = (X_train.shape[1],1)))

model.add(BatchNormalization())

cnn_model.add(Flatten())

cnn_model.add(Dense(units = 256, activation=LeakyReLU(alpha=0.001)))

cnn_model.add(Dense(units = 512, activation=LeakyReLU(alpha=0.001)))

cnn_model.add(Dense(units = 1, activation='sigmoid'))

cnn_model.compile(optimizer=RMSprop(lr=0.003), loss = 'binary_crossentropy', metrics=['accuracy'])
cnn_model.summary()
cnn_model.fit(X_train, 
              y_train, 
              epochs=10, 
              batch_size=1, 
              validation_data=(X_test, y_test),
             )


## 2 dim

X = np.empty((n_rows, S, N))
for i in range(0, n_rows):
    data = df.iloc[i]['data']
    X[i, :] = data

X_train, X_test, y_train, y_test = train_test_split(X, Y, test_size=0.33, random_state=42)

X_train = X_train.reshape([X_train.shape[0], 1, X_train.shape[2], X_train.shape[1]])
X_test = X_test.reshape([X_test.shape[0], 1, X_test.shape[2], X_test.shape[1]])

y_train = np.asarray(y_train).astype('float32').reshape((-1,1))
y_test = np.asarray(y_test).astype('float32').reshape((-1,1))

model = Sequential()

model.add(Conv3D(64, (1, 10, 5),  
                 padding='same', 
                 input_shape=[*X_train[0].shape, 1],
                 activation=LeakyReLU(alpha=0.001)))
model.add(MaxPooling3D((1, 96, 1)))
#model.add(Conv2D(256, (5, 20), kernel_regularizer=l2(0.0001), padding='same', activation='relu'))
#model.add(MaxPooling2D(5, ))
model.add(Flatten())
model.add(Dense(256, kernel_regularizer=l2(0.0001), activation=LeakyReLU(alpha=0.001)))
model.add(Dense(1, activation='softmax'))
model.summary()

model.compile(optimizer=RMSprop(lr=0.001), loss = 'binary_crossentropy', metrics=['accuracy'])
model.fit(X_train, y_train,
          batch_size=10,
          epochs=10,
          verbose=1,
         validation_data=(X_test, y_test))


X_train[0].shape
