==> scripts/7. Shion.ipynb <==
!pip install mlarray

import os

import numpy as np

import matlab

import pandas as pd

from scipy import io

df = pd.read_pickle('../workdata/aphrodite/sources_china.pkl')

df

AF_rows = df[df.label == 'AF']
items = { 'data': AF_rows.data.values, 'id': AF_rows.id.values }
io.savemat('AF.mat', items)

SR_rows = df[df.label == 'SR']
items = { 'data': SR_rows.data.values, 'id': SR_rows.id.values }
io.savemat('SR.mat', items)

print(data.values.shape)

from scipy import io

for i, row in df.iterrows():
    data = {
        'data': row,
        'label'
    }
    io.savemat('test.mat',data)
    break

type(file[0][0)

files = os.listdir('../workdata/shura/experiment2/rank_5/')

files
